challenge-5-ads-chatbot/requirements.txt
streamlit==1.44.1
vertexai==1.71.1
google-cloud-bigquery==3.34.0
google-genai==1.20.0